+++ source/ProcessManager.Orchestrations/Processes/BRS_025/V1/Model/RequestMeasurementsBusinessValidatedDto.cs
@@ -12,18 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Abstractions.Core.ValueObjects;
 using Energinet.DataHub.ProcessManager.Components.Abstractions.BusinessValidation;
-using Energinet.DataHub.ProcessManager.Components.Abstractions.EnqueueActorMessages;
-using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
+using Energinet.DataHub.ProcessManager.Components.MeteringPointMasterData.Model;
+using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_025.V1.Model;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_024.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_025.V1.Model;
 
-public record RequestYearlyMeasurementsRejectV1(
-    string OriginalActorMessageId,
-    string OriginalTransactionId,
-    ActorNumber ActorNumber,
-    ActorRole ActorRole,
-    string MeteringPointId,
-    List<ValidationErrorDto> ValidationErrors)
-        : IEnqueueRejectedDataDto;
+public record RequestMeasurementsBusinessValidatedDto(
+    RequestMeasurementsInputV1 Input,
+    IReadOnlyCollection<MeteringPointMasterData> HistoricalMeteringPointMasterData) : IBusinessValidatedDto;